requirements.txt
pillow
pdf2image
pytesseract
python-docx
pandas
